implemented feedback after code review
