+++ src/authn/SolidAuthnLogic.ts
@@ -2,12 +2,8 @@ import { namedNode, NamedNode, sym } from "rdflib";
 import { appContext, offlineTestID } from "./authUtil";
 import * as debug from '../util/debug'
 import { Session } from "@inrupt/solid-client-authn-browser";
-import { AuthenticationContext, AuthnLogic } from "../types";
-
-/**
- * Implements AuthnLogic relying on solid-auth-client
- */
-export class SolidAuthnLogic implements AuthnLogic {
+import { AuthenticationContext } from "../types";
+export class SolidAuthnLogic {
   private session: Session;
 
   constructor(solidAuthSession: Session) {
